move util.ts to graderv2
