+++ frontend/www/js/omegaup/components/arena/RunSubmitPopup.vue
@@ -88,7 +88,7 @@ import {
   LanguageInfo,
   supportedExtensions,
   supportedLanguages,
-} from '../../grader/util';
+} from '../../graderv2/util';
 @Component({
   components: {
     'omegaup-arena-code-view': arena_CodeView,
